ranking min_max partitioned into separate files, code inspected, python 3 improvements
