+++ golem/ranking/helper/ranking_stats.py
@@ -1,0 +1,6 @@
+class RankingStats(object):
+    computed = "computed"
+    wrong_computed = "wrong_computed"
+    requested = "requested"
+    payment = "payment"
+    resource = "resource"
